Spells refactor, 4th iteration
