+++ src/Engine/Objects/Player.cpp
@@ -1957,112 +1957,112 @@ int Player::ReceiveSpecialAttackEffect(
             case SPECIAL_ATTACK_CURSE:
                 SetCondition(Condition_Cursed, 1);
                 pAudioPlayer->PlaySound(SOUND_star1, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_WEAK:
                 SetCondition(Condition_Weak, 1);
                 pAudioPlayer->PlaySound(SOUND_star1, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_SLEEP:
                 SetCondition(Condition_Sleep, 1);
                 pAudioPlayer->PlaySound(SOUND_star1, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_DRUNK:
                 SetCondition(Condition_Drunk, 1);
                 pAudioPlayer->PlaySound(SOUND_star1, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_INSANE:
                 SetCondition(Condition_Insane, 1);
                 pAudioPlayer->PlaySound(SOUND_star4, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_POISON_WEAK:
                 SetCondition(Condition_Poison_Weak, 1);
                 pAudioPlayer->PlaySound(SOUND_star2, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_POISON_MEDIUM:
                 SetCondition(Condition_Poison_Medium, 1);
                 pAudioPlayer->PlaySound(SOUND_star2, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_POISON_SEVERE:
                 SetCondition(Condition_Poison_Severe, 1);
                 pAudioPlayer->PlaySound(SOUND_star2, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_DISEASE_WEAK:
                 SetCondition(Condition_Disease_Weak, 1);
                 pAudioPlayer->PlaySound(SOUND_star2, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_DISEASE_MEDIUM:
                 SetCondition(Condition_Disease_Medium, 1);
                 pAudioPlayer->PlaySound(SOUND_star2, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_DISEASE_SEVERE:
                 SetCondition(Condition_Disease_Severe, 1);
                 pAudioPlayer->PlaySound(SOUND_star2, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_PARALYZED:
                 SetCondition(Condition_Paralyzed, 1);
                 pAudioPlayer->PlaySound(SOUND_star4, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_UNCONSCIOUS:
                 SetCondition(Condition_Unconscious, 1);
                 pAudioPlayer->PlaySound(SOUND_star4, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_DEAD:
                 SetCondition(Condition_Dead, 1);
                 pAudioPlayer->PlaySound(SOUND_eradicate, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_PETRIFIED:
                 SetCondition(Condition_Petrified, 1);
                 pAudioPlayer->PlaySound(SOUND_eradicate, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_ERADICATED:
                 SetCondition(Condition_Eradicated, 1);
                 pAudioPlayer->PlaySound(SOUND_eradicate, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
@@ -2074,7 +2074,7 @@ int Player::ReceiveSpecialAttackEffect(
                     itemtobreak->SetBroken();
                     pAudioPlayer->PlaySound(SOUND_metal_vs_metal03h, 0, 0, -1, 0, 0);
                 }
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
@@ -2084,7 +2084,7 @@ int Player::ReceiveSpecialAttackEffect(
                 if (pActor->ActorHasItems[0].uItemID != ITEM_NULL) {
                     actoritems = &pActor->ActorHasItems[1];
                     if (pActor->ActorHasItems[1].uItemID != ITEM_NULL) {
-                        spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                        spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                         return 1;
                     }
                 }
@@ -2095,7 +2095,7 @@ int Player::ReceiveSpecialAttackEffect(
                 RemoveItemAtInventoryIndex(itemtostealinvindex);
                 pAudioPlayer->PlaySound(SOUND_metal_vs_metal03h, 0, 0, -1,
                                         0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
             }
@@ -2104,7 +2104,7 @@ int Player::ReceiveSpecialAttackEffect(
                 PlaySound(SPEECH_Aging, 0);
                 ++this->sAgeModifier;
                 pAudioPlayer->PlaySound(SOUND_eleccircle, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
@@ -2112,14 +2112,14 @@ int Player::ReceiveSpecialAttackEffect(
                 PlaySound(SPEECH_SPDrained, 0);
                 this->sMana = 0;
                 pAudioPlayer->PlaySound(SOUND_eleccircle, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
             case SPECIAL_ATTACK_FEAR:
                 SetCondition(Condition_Fear, 1);
                 pAudioPlayer->PlaySound(SOUND_star1, 0, 0, -1, 0, 0);
-                spell_fx_renderer->SetPlayerBuffAnim(0x99u, whichplayer);
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_DISEASE, whichplayer);
                 return 1;
                 break;
 
@@ -5071,7 +5071,7 @@ void Player::SetVariable(VariableType var_type, signed int var_value) {
         case VAR_QBits_QuestsDone:
             if (!_449B57_test_bit(pParty->_quest_bits, var_value) && pQuestTable[var_value]) {
                 bFlashQuestBook = 1;
-                spell_fx_renderer->SetPlayerBuffAnim(0x96u, GetPlayerIndex());
+                spell_fx_renderer->SetPlayerBuffAnim(BECOME_MAGIC_GUILD_MEMBER, GetPlayerIndex());
                 PlayAwardSound();
                 this->PlaySound(SPEECH_QuestGot, 0);
             }
@@ -5326,7 +5326,7 @@ void Player::SetVariable(VariableType var_type, signed int var_value) {
         case VAR_AutoNotes:
             if (!_449B57_test_bit(pParty->_autonote_bits, var_value) &&
                 pAutonoteTxt[var_value - 1].pText) {
-                spell_fx_renderer->SetPlayerBuffAnim(0x96u, GetPlayerIndex());
+                spell_fx_renderer->SetPlayerBuffAnim(BECOME_MAGIC_GUILD_MEMBER, GetPlayerIndex());
                 this->PlaySound(SPEECH_AwardGot, 0);
                 bFlashAutonotesBook = 1;
                 _506568_autonote_type = pAutonoteTxt[var_value - 1].eType;  // dword_72371C[2 * a3];
@@ -5520,7 +5520,7 @@ void Player::PlayAwardSound() {
 //----- (new function) --------------------------------------------------------
 void Player::PlayAwardSound_Anim() {
     int playerIndex = GetPlayerIndex();
-    spell_fx_renderer->SetPlayerBuffAnim(0x96u, playerIndex);
+    spell_fx_renderer->SetPlayerBuffAnim(BECOME_MAGIC_GUILD_MEMBER, playerIndex);
     PlayAwardSound();
 }
 
@@ -5540,7 +5540,7 @@ void Player::SetSkillByEvent(uint16_t Player::*skillToSet,
         this->*skillToSet = skillValue | (currSkillValue & 0xC0);
     }
     int playerIndex = GetPlayerIndex();
-    spell_fx_renderer->SetPlayerBuffAnim(0x96u, playerIndex);
+    spell_fx_renderer->SetPlayerBuffAnim(BECOME_MAGIC_GUILD_MEMBER, playerIndex);
     PlayAwardSound();
 }
 
@@ -5903,7 +5903,7 @@ void Player::AddVariable(VariableType var_type, signed int val) {
                 this->PlaySound(SPEECH_AwardGot, 0);
                 bFlashAutonotesBook = 1;
                 _506568_autonote_type = pAutonoteTxt[val].eType;
-                spell_fx_renderer->SetPlayerBuffAnim(0x97u, GetPlayerIndex());
+                spell_fx_renderer->SetPlayerBuffAnim(SPELL_QUEST_COMPLETED, GetPlayerIndex());
             }
             _449B7E_toggle_bit(pParty->_autonote_bits, val, 1);
             PlayAwardSound();
@@ -6072,7 +6072,7 @@ void Player::AddVariable(VariableType var_type, signed int val) {
 //----- (new function) --------------------------------------------------------
 void Player::PlayAwardSound_Anim97() {
     int playerIndex = GetPlayerIndex();
-    spell_fx_renderer->SetPlayerBuffAnim(0x97u, playerIndex);
+    spell_fx_renderer->SetPlayerBuffAnim(SPELL_QUEST_COMPLETED, playerIndex);
     PlayAwardSound();
 }
 
@@ -6646,7 +6646,7 @@ void Player::SubtractVariable(VariableType VarNum, signed int pValue) {
 //----- (new function) --------------------------------------------------------
 void Player::PlayAwardSound_Anim98() {
     int playerIndex = GetPlayerIndex();
-    spell_fx_renderer->SetPlayerBuffAnim(0x98u, playerIndex);
+    spell_fx_renderer->SetPlayerBuffAnim(SPELL_152, playerIndex);
     PlayAwardSound();
 }
 
@@ -7191,7 +7191,7 @@ void Player::OnInventoryLeftClick() {
 
         if (inventoryYCoord >= 0 && inventoryYCoord < INVENTORY_SLOTS_HEIGHT &&
             inventoryXCoord >= 0 && inventoryXCoord < INVENTORY_SLOTS_WIDTH) {
-            if (_50C9A0_IsEnchantingInProgress) {
+            if (IsEnchantingInProgress) {
                 unsigned int enchantedItemPos = this->GetItemListAtInventoryIndex(invMatrixIndex);
 
                 if (enchantedItemPos) {
@@ -7209,14 +7209,14 @@ void Player::OnInventoryLeftClick() {
                     pSpellInfo->spell_target_pid = enchantedItemPos - 1;
                     pSpellInfo->field_6 = this->GetItemMainInventoryIndex(invMatrixIndex);
                     ptr_50C9A4_ItemToEnchant = &this->pInventoryItemList[enchantedItemPos - 1];
-                    _50C9A0_IsEnchantingInProgress = 0;
+                    IsEnchantingInProgress = false;
 
                     pMessageQueue_50CBD0->Flush();
 
                     mouse->SetCursorImage("MICON1");
-                    _50C9D0_AfterEnchClickEventId = 113;
-                    _50C9D4_AfterEnchClickEventSecondParam = 0;
-                    _50C9D8_AfterEnchClickEventTimeout = 256;
+                    AfterEnchClickEventId = UIMSG_Escape;
+                    AfterEnchClickEventSecondParam = 0;
+                    AfterEnchClickEventTimeout = GameTime::FromSeconds(2);
                 }
 
                 return;
